+++ src/occlknndis.jl
@@ -15,34 +15,67 @@ end
         scal = false)
 One-class classification using "local" k-nearest neighbors distances.
 
-* `X` : X-data.
-* `nlv` : Nb. components for PCA.
-* `nsamp` : Nb. of training observations (rows of `X`) used to compute the 
-    empirical distribution of outlierness.
+* `X` : X-data (training).
+* `nlv` : Nb. PCA components for the distances computations.
+* `nsamp` : Nb. of observations (rows) sampled in the training `X`
+    used to compute the "H0" empirical distribution of outlierness.
 * `k` : Nb. of neighbors used to compute the outlierness.
-* `typc` : Type of cutoff ("mad" or "q"). See below.
-* `cri` : When `typc = "mad"`, constant used for computing the 
+* `typc` : Type of cutoff ("mad" or "q"), see below.
+* `cri` : When `typc = "mad"`, constant used to compute the 
     cutoff detecting extreme values.
-* `alpha` : When `typc = "q"`, risk-I level used for computing the cutoff 
-    detecting extreme values.
+* `alpha` : When `typc = "q"`, risk-I probability level to compute 
+    the cutoff detecting extreme values.
 * `scal` : Boolean. If `true`, each column of `X` is scaled
     by its uncorrected standard deviation.
 
-In this method, the "outlierness measure" `d` of a given observatio "q"  
-is defined as follows: 
-* The median of the distances between "q" and its `k` nearest neighbors
-in the training is computed, say d_q. 
-* Then, this same median distance is computed for each of the `k` nearest
-neighbors of "q" (i.e. each of the `k` nearest neighbors of "q" becomes 
-successively an observation "q" from which is computed the median distance to its
-`k` neighbors). This returns `k` median distances. Finally, the median of these
-`k` median distances is computed, and say d_q_nn.
-* The outlierness of "q" is defined by d_q / d_q_nn.  
 
-The distances are computed as Mahalanobis distances in a PCA score space 
-(internally computed; see argument `nlv`).
+Let us note q a given observation, and o[q] a neighbor of q 
+within the training data `X`. The `k` nearest neighbors of q
+define the neighborhood NNk(q) = {o.1[q], ...., o.k[q]} 
+(if q belongs to the training `X`, q is removed from NNk(q)). 
 
-See `?occknndis` for the cutoff computation (the same principle is applied). 
+The median distance of any observation q to its neighborhood NNk(q) is
+dk(q) = median{d(q, o.j[q]), j = 1,...,k}, referred to as global
+outlerness in function `occknndis`.
+
+The local outlierness of any observation q relatively to `X`, 
+say ldk(q), is computed as follows:
+* NNk(q) is selected and dk(q) is computed.
+* For each neighbor o.j[q] in NNk(q), global outlierness dk(o.j[q]) is 
+    computed. This returns the vector {dk(o.j[q]), j = 1,...,k}.
+* The local outlierness of q is computed as 
+    dk(q) / median{dk(o.j[q]), j = 1,...,k}.
+
+Outlierness ldk(q) is then compared to the outlierness distribution
+estimated for the training data `X`, say distribution H0.   
+If ldk(q) is extreme compared to H0, observation q may come from a 
+different distribution than the training data `X`.
+
+H0 is estimated by Monte Carlo, as follows:
+* A number of `nsamp` observations (rows) are sampled without replacement 
+    within the training data `X`.
+* For each of these `nsamp` training observations, say q.j {j = 1,...,nsamp},
+    outlierness ldk(q.j) is computed. This returns a vector of `nsamp` 
+    outlierness values {ldk(q.1), j = 1,...,nsamp}. 
+* This vector defines the empirical outlierness distribution of 
+    observations assumed to come from the same distribution as 
+    the training data `X` ("hypothesis H0"). 
+
+Then, function `predict` computes outlierness ldk(q) for each 
+new observation q. 
+
+In the function, distances are computed as Mahalanobis distances in a 
+PCA score space (internally computed), cf. argument `nlv`.
+
+The heuristic cutoff for detecting an extreme outlierness is computed
+as follows:
+* If `typc = "mad"`: cutoff = median(`d`) + `cri` * mad(`d`). 
+* If `typc = "q"`: the cutoff is estimated from the empirical cdf 
+    of `d`, depending on p-value `alpha`. 
+Output `dstand' is the outlierness standardized to the cutoff.
+
+Other details are the same as in `?occsd`. 
+
 
 ## Examples
 ```julia
@@ -82,8 +115,8 @@ plotxy(T[:, i], T[:, i + 1], group;
 
 #### End data
 
-nlv = 30 ; k = 5
-nsamp = 50
+nlv = 30
+k = 5 ; nsamp = 50
 fm = Jchemo.occlknndis(zXtrain; nsamp = nsamp,
     nlv = nlv, k = k, typc = "mad") ;
 fm.d
@@ -109,23 +142,21 @@ function occlknndis(X;
     X = ensure_mat(X)
     n = nro(X)
     fm = pcasvd(X; nlv = nlv, scal = scal)
-    T = fm.T
-    tscales = colstd(T)
-    scale!(T, tscales)
-    zn = collect(1:n)
-    samp = sample(zn, nsamp; replace = false)
+    tscales = colstd(fm.T)
+    scale!(fm.T, tscales)
+    samp = sample(1:n, nsamp; replace = false)
     ind = Int64.(zeros(k))
     d = zeros(nsamp)
     zd = zeros(nsamp)
     vd = zeros(k)
     @inbounds for i = 1:nsamp
-        # view vrow(T, i) does not work with getknn
-        res = getknn(rmrow(T, samp[i]), T[i:i, :]; k = k, metric = "eucl")
+        # view vrow(fm.T, i) does not work with getknn
+        res = getknn(rmrow(fm.T, samp[i]), fm.T[i:i, :]; k = k, metric = "eucl")
         ind .= res.ind[1]
         d[i] = median(res.d[1])
         for j = 1:k
             zind = ind[j]
-            zres = getknn(rmrow(T, zind), T[zind:zind, :]; k = k, metric = "eucl")
+            zres = getknn(rmrow(fm.T, zind), fm.T[zind:zind, :]; k = k, metric = "eucl")
             vd[j] = median(zres.d[1])
         end
         zd[i] = median(vd)
@@ -136,7 +167,7 @@ function occlknndis(X;
     e_cdf = StatsBase.ecdf(d)
     pval = 1 .- e_cdf(d)
     d = DataFrame(d = d, dstand = d / cutoff, pval = pval)
-    Occlknndis(d, fm, T, tscales, k, e_cdf, cutoff)
+    Occlknndis(d, fm, fm.T, tscales, k, e_cdf, cutoff)
 end
 
 function predict(object::Occlknndis, X)